requirements.txt
certifi==2020.6.20
click==8.0.3
colorama==0.4.4
colorharmonies==1.0.5
convcolors==1.0.0
dataclasses==0.8
extcolors==1.0.0
Flask==2.0.2
gunicorn==20.1.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.19.5
opencv-python-headless==4.5.5.62
Pillow==8.4.0
typing_extensions==4.0.1
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.6.0
